requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
cloudinary
django-cloudinary-storage
google-genai
firebase-admin
beautifulsoup4
requests
gunicorn
whitenoise
